☘️ refactor [#12.3.15]: 1차 개선 - GraphView 데이터 어댑터 분리 및 테스트 커버리지 확장
- `GraphView.tsx` 내부에 존재하던 데이터 변환 로직(`_adaptGraphData`)과 검증 헬퍼 함수들을 `graphViewAdapter.ts` 순수 유틸리티 모듈로 추출하여 UI 컴포넌트의 렌더링 관심사 분리
- Degree 기반 노드 Truncation 시 연결 수가 동일한 경우 ID 기반 사전순 정렬(Tie-break)을 적용하여 결정적(Deterministic)으로 동작하도록 개선
- `GraphView.test.ts`에 제한 개수 이하 시 잘림 없는 정상 경로(Happy path) 테스트 추가
- 모든 엣지가 유효하지 않아 Degree가 빈 경우에 대한 결정적 동작 검증 테스트 추가
- 컴포넌트 내부 사용되지 않는 모듈 린트 에러 정리

🔗 Related:
- Issue [#1048]
- @ai-bot-review (https://github.com/jjaayy2222/flownote-mvp/pull/1564#pullrequestreview-4443345587)
